Moving interfaces to a separate project
