--- UI/WebStore/Infrastructure/Services/InCookies/CookiesCartService.cs
@@ -1,117 +1,0 @@
-﻿using System.Linq;
-using Microsoft.AspNetCore.Http;
-using Newtonsoft.Json;
-using WebStore.Domain;
-using WebStore.Domain.Entities;
-using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
-using WebStore.Infrastructure.Mapping;
-
-namespace WebStore.Infrastructure.Services.InCookies
-{
-    public class CookiesCartService : ICartService
-    {
-        private readonly IProductData _ProductData;
-        private readonly IHttpContextAccessor _HttpContextAccessor;
-        private readonly string _CartName;
-
-        private Cart Cart
-        {
-            get
-            {
-                var context = _HttpContextAccessor.HttpContext;
-                var cookies = context.Response.Cookies;
-                var cart_cookies = context.Request.Cookies[_CartName];
-                if (cart_cookies is null)
-                {
-                    var cart = new Cart();
-                    cookies.Append(_CartName, JsonConvert.SerializeObject(cart));
-                    return cart;
-                }
-
-                ReplaceCookies(cookies, cart_cookies);
-                return JsonConvert.DeserializeObject<Cart>(cart_cookies);
-            }
-            set => ReplaceCookies(_HttpContextAccessor.HttpContext.Response.Cookies, JsonConvert.SerializeObject(value));
-        }
-
-        private void ReplaceCookies(IResponseCookies cookies, string cookie)
-        {
-            cookies.Delete(_CartName);
-            cookies.Append(_CartName, cookie);
-        }
-
-        public CookiesCartService(IProductData ProductData, IHttpContextAccessor HttpContextAccessor)
-        {
-            _ProductData = ProductData;
-            _HttpContextAccessor = HttpContextAccessor;
-
-            var user = HttpContextAccessor.HttpContext.User;
-            var user_name = user.Identity.IsAuthenticated ? $"[{user.Identity.Name}]" : null;
-            _CartName = $"WebStore.Cart{user_name}";
-        }
-
-        public void AddToCart(int id)
-        {
-            var cart = Cart;
-            var item = cart.Items.FirstOrDefault(i => i.ProductId == id);
-
-            if (item is null)
-                cart.Items.Add(new CartItem {ProductId = id, Quantity = 1});
-            else
-                item.Quantity++;
-
-            Cart = cart;
-        }
-
-        public void DecrementFromCart(int id)
-        {
-            var cart = Cart;
-            var item = cart.Items.FirstOrDefault(i => i.ProductId == id);
-            if(item is null) return;
-
-            if (item.Quantity > 0)
-                item.Quantity--;
-
-            if (item.Quantity == 0)
-                cart.Items.Remove(item);
-
-            Cart = cart;
-        }
-
-        public void RemoveFromCart(int id)
-        {
-            var cart = Cart;
-            var item = cart.Items.FirstOrDefault(i => i.ProductId == id);
-            if (item is null) return;
-
-            cart.Items.Remove(item);
-
-            Cart = cart;
-        }
-
-        public void Clear()
-        {
-            var cart = Cart;
-
-            cart.Items.Clear();
-
-            Cart = cart;
-        }
-
-        public CartViewModel TransformFromCart()
-        {
-            var products = _ProductData.GetProducts(new ProductFilter
-            {
-                Ids = Cart.Items.Select(item => item.ProductId).ToArray()
-            });
-
-            var products_view_models = products.ToView().ToDictionary(p => p.Id);
-
-            return new CartViewModel
-            {
-                Items = Cart.Items.Select(item => (products_view_models[item.ProductId], item.Quantity))
-            };
-        }
-    }
-}
