--- UI/WebStore/Infrastructure/Services/InSQL/SqlProductData.cs
@@ -1,46 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
-using WebStore.DAL.Context;
-using WebStore.Domain;
-using WebStore.Domain.Entities;
-using WebStore.Infrastructure.Interfaces;
-
-namespace WebStore.Infrastructure.Services.InSQL
-{
-    public class SqlProductData : IProductData
-    {
-        private readonly WebStoreDB _db;
-
-        public SqlProductData(WebStoreDB db) => _db = db;
-
-        public IEnumerable<Section> GetSections() => _db.Sections;
-
-        public IEnumerable<Brand> GetBrands() => _db.Brands;
-
-        public IEnumerable<Product> GetProducts(ProductFilter Filter = null)
-        {
-            IQueryable<Product> query = _db.Products
-               .Include(product => product.Brand)
-               .Include(product => product.Section);
-
-            if (Filter?.Ids?.Length > 0)
-                query = query.Where(product => Filter.Ids.Contains(product.Id));
-            else
-            {
-                if (Filter?.BrandId != null)
-                    query = query.Where(product => product.BrandId == Filter.BrandId);
-
-                if (Filter?.SectionId != null)
-                    query = query.Where(product => product.SectionId == Filter.SectionId);
-            }
-
-            return query/*.ToArray()*/;
-        }
-
-        public Product GetProductById(int id) => _db.Products
-           .Include(product => product.Brand)
-           .Include(product => product.Section)
-           .FirstOrDefault(product => product.Id == id);
-    }
-}
